chore: move a bunch of methods
