+++ package.json
@@ -76,12 +76,12 @@
     "@babel/preset-env": "^7.6.0",
     "@glimmer/compiler": "0.38.5-alpha.3",
     "@glimmer/env": "^0.1.7",
-    "@glimmer/interfaces": "0.38.5-alpha.3",
+    "@glimmer/interfaces": "0.39.1",
     "@glimmer/node": "0.38.5-alpha.3",
     "@glimmer/opcode-compiler": "0.38.5-alpha.3",
     "@glimmer/program": "0.38.5-alpha.3",
     "@glimmer/reference": "0.38.5-alpha.3",
-    "@glimmer/runtime": "0.38.5-alpha.3",
+    "@glimmer/runtime": "0.39.1",
     "@types/qunit": "^2.5.4",
     "@types/rsvp": "^4.0.3",
     "auto-dist-tag": "^1.0.0",
