+++ packages/@ember/-internals/glimmer/lib/component-managers/input.ts
@@ -2,9 +2,9 @@ import { ENV } from '@ember/-internals/environment';
 import { set } from '@ember/-internals/metal';
 import { Owner } from '@ember/-internals/owner';
 import { assert, debugFreeze } from '@ember/debug';
-import { ComponentCapabilities, Dict } from '@glimmer/interfaces';
+import { ComponentCapabilities, Dict, DynamicScope, VMArguments as Arguments } from '@glimmer/interfaces';
 import { CONSTANT_TAG, createTag, isConst, VersionedPathReference } from '@glimmer/reference';
-import { Arguments, Bounds, DynamicScope, PreparedArguments } from '@glimmer/runtime';
+import { Bounds, PreparedArguments } from '@glimmer/runtime';
 import { Destroyable } from '@glimmer/util';
 import Environment from '../environment';
 import { RootReference } from '../utils/references';
