+++ packages/@ember/-internals/glimmer/lib/component-managers/outlet.ts
@@ -6,10 +6,9 @@ import EngineInstance from '@ember/engine/instance';
 import { _instrumentStart } from '@ember/instrumentation';
 import { assign } from '@ember/polyfills';
 import { DEBUG } from '@glimmer/env';
-import { ComponentCapabilities, Option, Simple } from '@glimmer/interfaces';
+import { ComponentCapabilities, Option, Simple, VMArguments as Arguments } from '@glimmer/interfaces';
 import { CONSTANT_TAG, createTag, Tag, VersionedPathReference } from '@glimmer/reference';
 import {
-  Arguments,
   Bounds,
   ComponentDefinition,
   ElementOperations,
