+++ packages/@ember/-internals/glimmer/lib/component-managers/root.ts
@@ -3,8 +3,8 @@ import { ENV } from '@ember/-internals/environment';
 import { Factory } from '@ember/-internals/owner';
 import { _instrumentStart } from '@ember/instrumentation';
 import { DEBUG } from '@glimmer/env';
-import { ComponentCapabilities, Option } from '@glimmer/interfaces';
-import { Arguments, ComponentDefinition, EMPTY_ARGS } from '@glimmer/runtime';
+import { ComponentCapabilities, Option, VMArguments as Arguments } from '@glimmer/interfaces';
+import { ComponentDefinition, EMPTY_ARGS } from '@glimmer/runtime';
 import { DIRTY_TAG } from '../component';
 import Environment from '../environment';
 import { DynamicScope } from '../renderer';
