+++ packages/@ember/-internals/glimmer/lib/helpers/-normalize-class.ts
@@ -1,5 +1,5 @@
 import { dasherize } from '@ember/string';
-import { Arguments, VM } from '@glimmer/runtime';
+import { VMArguments as Arguments, VM } from '@glimmer/interfaces';
 import { InternalHelperReference } from '../utils/references';
 
 function normalizeClass({ positional }: any) {
