+++ packages/@ember/-internals/glimmer/lib/helpers/mut.ts
@@ -3,7 +3,7 @@
 */
 import { symbol } from '@ember/-internals/utils';
 import { assert } from '@ember/debug';
-import { Arguments, VM } from '@glimmer/runtime';
+import { VMArguments as Arguments, VM } from '@glimmer/interfaces';
 import { INVOKE, UPDATE } from '../utils/references';
 
 /**
